pyproject.toml
[project]
name = "c3-chat-py"
version = "0.1.0"
description = "Chat application with Claude API and GitHub integration"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.49.0",
    "chainlit>=2.4.1",
    "composio-claude>=0.7.8",
    "composio-core>=0.7.8",
    "composio-openai>=0.7.8",
    "composio-openai-agents>=0.7.8",
    "python-dotenv>=1.0.1",
]

[dependency-groups]
dev = [
    "ipykernel>=6.29.5",
    "ipywidgets>=8.1.5",
]

[tool.setuptools]
packages = ["src"]

[tool.setuptools.package-data]
"*" = ["py.typed"]
